Refactoring - moved logic to LinkNodeVisitor
