+++ README.md
@@ -7,4 +7,3 @@ Your code has more knowledge then it meets the eye.
 
 * develop on PHP 8.1 for the best contributor experience
 * downgrade on release to PHP 7.2, so anyone can use them
-
