+++ packages/latte-phpstan-compiler/src/PhpParser/NodeVisitor/LinkNodeVisitor.php
@@ -17,16 +17,45 @@ use Reveal\LattePHPStanCompiler\Contract\LatteToPhpCompilerNodeVisitorInterface;
 use Reveal\LattePHPStanCompiler\Contract\LinkProcessorInterface;
 use Reveal\LattePHPStanCompiler\Exception\LattePHPStanCompilerException;
 use Reveal\LattePHPStanCompiler\LinkProcessor\LinkProcessorFactory;
+use Reveal\LattePHPStanCompiler\Nette\PresenterFactoryFaker;
+use Reveal\TemplatePHPStanCompiler\ValueObject\VariableAndType;
 use Symplify\Astral\Naming\SimpleNameResolver;
 use Symplify\Astral\NodeValue\NodeValueResolver;
 
 final class LinkNodeVisitor extends NodeVisitorAbstract implements LatteToPhpCompilerNodeVisitorInterface
 {
+    /**
+     * @var VariableAndType[]
+     */
+    private array $variablesAndTypes = [];
+
+    private ?VariableAndType $actualclass = null;
+
     public function __construct(
         private SimpleNameResolver $simpleNameResolver,
         private NodeValueResolver $nodeValueResolver,
         private LinkProcessorFactory $linkProcessorFactory,
+        private PresenterFactoryFaker $presenterFactoryFaker,
     ) {
+    }
+
+    /**
+     * @param VariableAndType[] $variablesAndTypes
+     */
+    public function setVariablesAndTypes(array $variablesAndTypes): void
+    {
+        $this->variablesAndTypes = $variablesAndTypes;
+    }
+
+    public function beforeTraverse(array $nodes)
+    {
+        // finding $actualClass
+        foreach ($this->variablesAndTypes as $variableAndType) {
+            if ($variableAndType->getVariable() === 'actualClass') {
+                $this->actualclass = $variableAndType;
+            }
+        }
+        return null;
     }
 
     /**
@@ -75,8 +104,7 @@ final class LinkNodeVisitor extends NodeVisitorAbstract implements LatteToPhpCom
             throw new LattePHPStanCompilerException();
         }
 
-        $targetName = ltrim($targetName, '/');
-
+        $targetName = $this->remapTarget($targetName);
         $linkProcessor = $this->linkProcessorFactory->create($targetName);
         if (! $linkProcessor instanceof LinkProcessorInterface) {
             return null;
@@ -124,5 +152,43 @@ final class LinkNodeVisitor extends NodeVisitorAbstract implements LatteToPhpCom
         }
 
         return $linkParams;
+    }
+
+    private function remapTarget(string $targetName): string
+    {
+        $targetName = ltrim($targetName, '/:');
+
+        if (str_ends_with($targetName, '!')) {
+            return $targetName;
+        }
+
+        $targetNameParts = explode(':', $targetName);
+        if (count($targetNameParts) === 3) {
+            return $targetName;
+        }
+
+        if ($this->actualclass === null) {
+            return $targetName;
+        }
+
+        $presenterFactory = $this->presenterFactoryFaker->getPresenterFactory();
+        $presenterName = @$presenterFactory->unformatPresenterClass($this->actualclass->getTypeAsString());
+
+        if ($presenterName === null) {
+            return $targetName;
+        }
+
+        if (count($targetNameParts) === 1) {
+            return $presenterName . ':' . $targetName;
+        }
+
+        $presenterNameParts = explode(':', $presenterName, 2);
+
+        $module = isset($presenterNameParts[1]) ? $presenterNameParts[0] : null;
+        if ($module) {
+            return $module . ':' . $targetName;
+        }
+
+        return $targetName;
     }
 }
